Module ruma_appservice_api::thirdparty

Endpoints for third party lookups

Modules

get_location_for_protocol
Endpoint to retrieve a list of Matrix portal rooms that lead to the matched third party location.

get_location_for_room_alias
Endpoint to retrieve an array of third party network locations from a Matrix room alias.

get_protocol
Endpoint to present clients with specific information about the various third party networks that an application service supports.

get_user_for_protocol
Endpoint to retrieve a Matrix User ID linked to a user on the third party network, given a set of user parameters.

get_user_for_user_id
Endpoint to retrieve an array of third party users from a Matrix User ID.